refactor(ui): extract markdown render owners
